+++ src/Main/UIW_Settings.cpp
@@ -3,9 +3,9 @@
 // * Description:    UI Wizard Settings Source File                          * //
 // * Author:         TT                                                      * //
 // * Website:        https://github.com/The-Wizardium/UI-Wizard              * //
-// * Version:        0.2.0                                                   * //
+// * Version:        0.2.1                                                   * //
 // * Dev. started:   12-12-2024                                              * //
-// * Last change:    16-09-2025                                              * //
+// * Last change:    19-09-2025                                              * //
 /////////////////////////////////////////////////////////////////////////////////
 
 
@@ -25,6 +25,7 @@ void UIWizardSettings::ResetSettings() {
 
 	windowBgColor = windowBgColorDefault;
 	windowState = windowStateDefault;
+	windowStatePrevious = windowStatePreviousDefault;
 
 	frameStylePrevious = frameStylePreviousDefault;
 	frameStyle = frameStyleDefault;